pyproject.toml
[project]
name = "geo-resolver"
version = "0.1.0"
description = "Natural language to GeoJSON boundary resolver"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=1.1.0",
    "shapely>=2.0",
    "pyproj>=3.6",
    "openai>=1.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
api = ["fastapi>=0.110", "uvicorn[standard]>=0.27", "sse-starlette>=2.0"]
dev = ["pytest", "httpx", "ipython"]

[project.scripts]
geo-resolve = "geo_resolver.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["geo_resolver*"]
